update store to be reactive based on tags, selected by tab
